Fixed little errors that Kayla helped on in ListItem

+++ ListItem.java
@@ -1,8 +1,8 @@
 import java.util.function.*;
 public class ListItem {
 	private LongConsumer item;
-	private int ord;
-	public ListItem(LongConsumer item, int ord)
+	private long ord;
+	public ListItem(LongConsumer item, long ord)
 	{
 		this.item = item;
 		this.ord = ord;
@@ -11,7 +11,7 @@ public class ListItem {
 	{
 		return this.item;
 	}
-	public int getOrd()
+	public long getOrd()
 	{
 		return this.ord;
 	}
